Create README.md
Created README,md for PARK-Tokens to allow Modal to :100: 

This contract allows the sale of parking tokens, the creation of parking spot NFTs, the ability to redeem them for parking hours, sell NFTs at a higher exchange rate, and lets the contract owner withdraw Ether. This forms the basis of an interactive NFT-based parking solution!